Fix remaining bug with unsigned comparisons.
Part of the problem is that some unsigned comparison conditions (_U{LT,GE}) are not decoded properly. However, the main problem is that the semantics of the carry flag in subtraction seem to have changed since the earlier version of FRISC.

In particular, it used to be the case that the carry flag took the value of the borrow from the MSB, but now takes the value of the carry from the MSB even for subtraction. To make that work out properly with the SBC instruction, its semantics were changed from

SBC A, B = A - B - C

to

SBC A, B = A - B + C

I've hopefully updated all the necessary pieces of code to address all of this.

I'm also adding some "end-to-end" tests for this, as it is quite difficult to debug with the existing test suite. The new tests take a program written in FRISC assembly, assemble it, run it, and check some result.

I've also run grunt in the end, without lint as there are a lot of lint errors in existing code.